# native/application/api/openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_gcc798_lightning_application_api_domain_request.CleanOperLogRequest:
    properties:
      days:
        description: 清理多少天之前的日志
        minimum: 1
        type: integer
    required:
    - days
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.CreateOperLogRequest:
    properties:
      businessType:
        description: 业务类型
        type: string
      costTime:
        description: 耗时（毫秒）
        type: integer
      deviceType:
        description: 终端类型：web/ios/android/wechat
        type: string
      errorMsg:
        description: 错误信息
        type: string
      jsonResult:
        description: 返回结果
        type: string
      method:
        description: 调用方法
        type: string
      operIp:
        description: 操作IP
        type: string
      operLocation:
        description: 操作地点
        type: string
      operName:
        description: 操作者
        type: string
      operParam:
        description: 请求参数
        type: string
      operUrl:
        description: 请求URL
        type: string
      requestMethod:
        description: 请求方式：GET/POST
        type: string
      status:
        description: 操作状态：0成功 1失败
        type: string
      title:
        description: 模块标题
        type: string
      userAgent:
        description: UA
        type: string
    required:
    - title
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.PageOperLogRequest:
    properties:
      businessType:
        description: 业务类型（可选）
        type: string
      endTime:
        description: 结束时间（可选）
        type: string
      isAsc:
        description: 'IsAsc 排序方向，asc或desc，支持多个，逗号分隔，如: "asc,desc"'
        type: string
      operName:
        description: 操作者（可选,模糊查询）
        type: string
      orderByColumn:
        description: 'OrderByColumn 排序列，支持多列排序，逗号分隔，如: "id,createTime"'
        type: string
      pageNum:
        description: PageNum 当前页码，从1开始
        type: integer
      pageSize:
        description: PageSize 每页数量
        type: integer
      startTime:
        description: 开始时间（可选）
        type: string
      status:
        description: 操作状态（可选,nil表示全部,"0"成功 "1"失败）
        type: string
      title:
        description: 模块标题（可选,模糊查询）
        type: string
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.UpdateOperLogRequest:
    properties:
      businessType:
        description: 业务类型
        type: string
      costTime:
        description: 耗时（毫秒）
        type: integer
      deviceType:
        description: 终端类型：web/ios/android/wechat
        type: string
      errorMsg:
        description: 错误信息
        type: string
      id:
        description: 日志ID
        type: integer
      jsonResult:
        description: 返回结果
        type: string
      method:
        description: 调用方法
        type: string
      operIp:
        description: 操作IP
        type: string
      operLocation:
        description: 操作地点
        type: string
      operName:
        description: 操作者
        type: string
      operParam:
        description: 请求参数
        type: string
      operUrl:
        description: 请求URL
        type: string
      requestMethod:
        description: 请求方式：GET/POST
        type: string
      status:
        description: 操作状态：0成功 1失败
        type: string
      title:
        description: 模块标题
        type: string
      userAgent:
        description: UA
        type: string
    required:
    - id
    type: object
...
  github_com_gcc798_lightning_application_api_domain_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: 操作成功
        type: string
    type: object
...
host: localhost:9009
info:
  contact:
    email: support@example.com
    name: 技术支持
    url: https://example.com/support
  description: Quick Admin RESTful API 接口文档，支持双 Token 认证机制
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: NTZ API 文档
  version: "1.0"
paths:
...
  /api/v1/operLog:
    post:
      consumes:
      - application/json
      description: 创建新的操作日志记录
      parameters:
      - description: 创建操作日志请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.CreateOperLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 创建操作日志
      tags:
      - 操作日志
    put:
      consumes:
      - application/json
      description: 更新操作日志记录
      parameters:
      - description: 更新操作日志请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.UpdateOperLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 更新操作日志
      tags:
      - 操作日志
...
  /api/v1/operLog/clean:
    post:
      consumes:
      - application/json
      description: 清理指定天数之前的操作日志
      parameters:
      - description: 清理日志请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.CleanOperLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 清理成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 清理操作日志
      tags:
      - 操作日志
  /api/v1/operLog/page:
    post:
      consumes:
      - application/json
      description: 分页查询操作日志列表，支持按标题、操作者、业务类型、状态、时间范围筛选
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.PageOperLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 分页查询操作日志列表
      tags:
      - 操作日志
...
securityDefinitions:
  Bearer:
    description: '格式: "Bearer {access_token}"，AccessToken 从登录接口获取'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...